Separate Clients, Automate test.
